% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmarking Different Accelerating Methods for Various Fixed-Point Iteration Problems}
\usage{
benchmark(
  task = "poissmix",
  algorithm = c("raw"),
  ntimes = 100,
  control = list(tol = 1e-07, convtype = "parameter", maxiter = 2000),
  control.spec = list(),
  verbose = TRUE,
  sample_every_time = TRUE,
  seed = NULL,
  store_tasks = FALSE,
  ...
)
}
\arguments{
\item{task}{A list or function or string for the problem to be benchmarked. Explained in Details below.}

\item{algorithm}{An string array of algorithms that will be run in the benchmarking. "raw" for the original algorithm in the task. "squarem" for SQUAREM. "daarem" for DAAREM. "qn" for Quasi-Newton, "pem" for parabolic-EM and "nes" for restart Nesterov.}

\item{ntimes}{Integer, number of repetitions for the task in benchmarking.}

\item{control}{A list for control parameters that will be shared to all the algorithm tried.}

\item{control.spec}{A named list to pass control parameters for specific algorithms. For example \code{list(squarem=listA, qn=listB)} will pass parameters in \code{listA} to SQUAREM and those in \code{listB} to Quasi-Newton.}

\item{verbose}{A bool value indicating whether to print out informating when the function is running.}

\item{sample_every_time}{A bool value indicating whether we should create a new task list in every repetition. This is only usable when \code{task} is a function.}

\item{seed}{An integer for random seed}

\item{store_tasks}{A bool value indicating whether to store all tasks in the result.}

\item{...}{Other arguments required by \code{task} if it is a function.}
}
\value{
A list of results in format of "benchmark" class.
 \item{result_table}{A dataframe containing necessary information for each repetition.}
 \item{all_results}{A list combining all results lists from all methods in all repetitions.}
 \item{all_tasks}{A list containing all task lists in every repetition. If \code{store_task} is \code{FALSE} this will be empty.}
 \item{task}{A string for task name.}
}
\description{
A benchmarking function used to test the performance of different accelerating methods on various fixed-point interation problems.
}
\details{
The \code{task} argument above indicates the problem to be benchmarked. In the simplest case, it will be a list containing \code{initfn} for parameter initialization function; \code{fixptfn} for fixed-point updating function; \code{objfn} for objective function and other arguments required for \code{fixptfn} and \code{objfn}.
         But \code{task} can also be a function taking parameters and return such list.
         We also register some character names for specific tasks: "poissmix" for Poisson mixture; "mvt_mmf" for Multivariate t-distribution; "lasso" for LASSO logistic regression; "bvs" for variational Bayes variable selection; "tsne" for tSNE in COIL-20 and "sinkhorn" for Sinkhorn iteration in matrix balancing.
}
\examples{
\dontrun{
set.seed(54321)
benchmark("poissmix", c("raw", "squarem", "daarem", "pem", "qn", "nes"),  ntimes=100)
}

}
