% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COIL_20-data.R
\docType{data}
\name{COIL_20}
\alias{COIL_20}
\title{COIL-20 objects data}
\format{
A List where \code{"$X"} is an (1440, 16384) matrix contains all raster images and \code{"$label"} is object label for each image.
}
\source{
\href{https://www.cs.columbia.edu/CAVE/software/softlib/coil-20.php}{COIL-20}
}
\usage{
data(COIL_20)
}
\description{
Data contains 1440 128x128 images for 20 objects picturing from different angles.
}
\examples{
data(COIL_20)
\donttest{image(matrix(COIL_20$X[1,], 128, 128), useRaster=TRUE, axes=FALSE, col=gray.colors(33))}

}
\references{
"Columbia Object Image Library (COIL-20)," S. A. Nene, S. K. Nayar and H. Murase, Technical Report CUCS-005-96, February 1996.
}
\keyword{datasets}
