% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_task.R
\name{lasso_task}
\alias{lasso_task}
\title{LASSO logistic regression problem}
\usage{
lasso_task(lam)
}
\arguments{
\item{lam}{Weight for the L1 penalty}
}
\value{
A list containing all components needed for benchmarking the problem
 \item{initfn}{Parameter random initializing function}
 \item{fixptfn}{Updating function for the fixed point iteration problem}
 \item{objfn}{Function calculating the objective value for current parameter}
 \item{...}{Other arguments required in functions above}
}
\description{
A function that will generate the task list LASSO logistic regression problem.
}
\examples{
\dontrun{
set.seed(54321)
problem = lasso_task(lam=1)
benchmark(
  problem,
  algorithm=c("raw", "squarem", "daarem", "pem", "qn", "nes"),
  ntimes = 200
)
}

}
\references{
Guyon I, Gunn SR, Ben-Hur A, Dror G (2004). Result analysis of the NIPS 2003 feature selection challenge. In: NIPS, volume 4, 545–552.
}
