% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Madelon-data.R
\docType{data}
\name{Madelon}
\alias{Madelon}
\title{Madelon data}
\format{
A List where \code{"$X"} is design matrix of size (2600, 500), \code{"$y"} is the binary outcome for each row.
}
\usage{
data(Madelon)
}
\description{
Simulated data with highly ill-conditioned design matrix.
}
\examples{
data(Madelon)
\donttest{table(Madelon$y)}

}
\references{
Guyon, Isabelle, et al. "Result analysis of the nips 2003 feature selection challenge." Advances in neural information processing systems 17 (2004).
(\href{https://proceedings.neurips.cc/paper/2004/file/5e751896e527c862bf67251a474b3819-Paper.pdf}{NeurIPS})
}
\keyword{datasets}
