% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt_task.R
\name{mvt_mmf_task}
\alias{mvt_mmf_task}
\title{Multivarite t distribution problem}
\usage{
mvt_mmf_task(df)
}
\arguments{
\item{df}{Degree of freedom for the t-distribution}
}
\value{
A list containing all components needed for benchmarking the problem
 \item{initfn}{Parameter random initializing function}
 \item{fixptfn}{Updating function for the fixed point iteration problem}
 \item{objfn}{Function calculating the objective value for current parameter}
 \item{...}{Other arguments required in functions above}
}
\description{
A function that will generate the task list for problem of estimating parameters of a multivariate t distribution.
We would simulate 1000 2-dimensional t distribution data with mean 0 and \eqn{\Sigma = diag(0.1,1)}.
And the base fixed point algorithm is MMF in Hasannasab (2021).
}
\examples{
\dontrun{
set.seed(54321)
problem = mvt_mmf_task(df=25)
benchmark(
  problem,
  algorithm=c("raw", "squarem", "daarem", "pem", "qn", "nes"),
  ntimes = 200
)
}

}
\references{
Hasannasab M, Hertrich J, Laus F, Steidl G (2021). Alternatives to the EM algorithm for ML estimation of location, scatter matrix, and degree of freedom of the Student t distribution. Numerical Algorithms, 87(1): 77–118.
}
