% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{plot.benchmark}
\alias{plot.benchmark}
\title{Plot Method for \code{benchmark} Class}
\usage{
\method{plot}{benchmark}(
  x,
  start_from = 1,
  min_loss = NULL,
  type = "iter",
  eps = 1e-10,
  xlog = TRUE,
  ylog = TRUE,
  ...
)
}
\arguments{
\item{x}{benchmark class output by benchmark function}

\item{start_from}{A integer indicating which repetition should be plot out. If that has NA values it will search the following repetitions and plot out the first one without NA.}

\item{min_loss}{A real number indicating the minimum loss that should be subtracted out when doing plots. If NULL, the minimum loss will be subtracted out.}

\item{type}{A string indicating the x-axis. "iter" for fpevals and "elapsed" for elapsed time.}

\item{eps}{A small positive real number setting a lower bound for the standardized loss. Useful when doing log scale plots.}

\item{xlog}{A bool value indicating whether x should be plotted in log scale}

\item{ylog}{A bool value indicating whether y should be plotted in log scale}

\item{...}{Other arguments for plot function.}
}
\description{
Plot Method for \code{benchmark} Class
}
\examples{
\dontrun{
set.seed(54321)
out = benchmark("poissmix", c("raw", "squarem", "daarem", "pem", "qn", "nes"),  ntimes=100)
plot(out)
}

}
