% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkhorn_task.R
\name{sinkhorn_task}
\alias{sinkhorn_task}
\title{Matrix balancing problem}
\usage{
sinkhorn_task(mat = "Marshall_Olkin", order = 20)
}
\arguments{
\item{mat}{Matrix to balancing. Can be a matrix or character name "Hessenberg" or "Marshall_Olkin" for specific matrices with those names.}

\item{order}{Order of the matrix. Only meaningful when \code{mat} argument is "Hessenberg"}
}
\value{
A list containing all components needed for benchmarking the problem
 \item{initfn}{Parameter random initializing function}
 \item{fixptfn}{Updating function for the fixed point iteration problem}
 \item{objfn}{Function calculating the objective value for current parameter}
 \item{...}{Other arguments required in functions above}
}
\description{
A function that will generate the task list that can be used to benchmark acceleration of Sinkhorn iteration in Matrix balancing problem
}
\examples{
\dontrun{
set.seed(54321)
problem = sinkhorn_task(mat="Hessenberg", order=25)
benchmark(
  problem,
  algorithm=c("raw", "squarem", "daarem", "pem", "qn", "nes"),
  ntimes = 200
)
}

}
\references{
Sinkhorn R, Knopp P (1967). Concerning nonnegative matrices and doubly stochastic matrices. Pacific Journal of Mathematics, 21(2): 343–348.

Parlett B, Landis TL (1982). Methods for scaling to doubly stochastic form. Linear Algebra and its Applications, 48: 53–79
}
