% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squarem.R
\name{squarem}
\alias{squarem}
\title{Squared Extrapolation Methods for Accelerating Slowly-Convergent Fixed-Point Iterations}
\usage{
squarem(par, fixptfn, objfn, ..., control = list())
}
\arguments{
\item{par}{Vector for initial parameters}

\item{fixptfn}{Fixed point updating function}

\item{objfn}{Objective function}

\item{...}{Other arguments required by \code{fixptfn} and \code{objfn}}

\item{control}{A list containing parameters controlling the algorithm}
}
\value{
A list of results
 \item{par}{Parameter values, x* that are the fixed-point of fixptfn F such that x*=F(x*) if convergence is successful.}
 \item{value.objfn}{The objective function value at termination.}
 \item{fpevals}{Number of times the fixed-point function \code{fixptfn} was evaluated.}
 \item{objfevals}{Number of times the objective function \code{objfn} was evaluated.}
 \item{iter}{Numbers of iteration used at termination. (for different algorithms, multiple fixed point iteration might be evaluated in one iteration)}
 \item{convergence}{An integer code indicating whether the algorithm converges. 1 for convergence and 0 denote failure.}
 \item{objfn.track}{An array tracking objective function values along the algorithm}
 \item{par.track}{A matrix tracking parameters along the algorithm, where each row is an array of parameters at some iteration. If not tracking paramters, this will be \code{NULL}}
}
\description{
Globally-convergent, partially monotone, acceleration schemes for accelerating the convergence of *any* smooth, monotone, slowly-converging contraction mapping. It can be used to accelerate the convergence of a wide variety of iterations including the expectation-maximization (EM) algorithms and its variants, majorization-minimization (MM) algorithm, power method for dominant eigenvalue-eigenvector, Google's page-rank algorithm, and multi-dimensional scaling.
This is a modification of the original squarem in \code{SQUAREM} package, including choice of projection and user defined convergence function.
}
\details{
The task it to \strong{minimize} \code{objfn}. Default values of \code{control} are: \code{method=3, step.min0=1, step.max0=1, mstep=4, objfn.inc=1, projection=function(x) x, tol=1e-7, maxiter=2000, convtype="parameter", par.track=FALSE, conv.spec=NULL}.
\describe{
 \item{method}{An integer variable that denotes the particular SQUAREM scheme to be used. Value should be either 1, 2, or 3. These correspond to the 3 schemes discussed in Varadhan and Roland (2008). Default is 3.}
 \item{step.min0}{A scalar denoting the minimum steplength taken by a SQUAREM algorithm. Default is 1. For contractive fixed-point iterations (e.g. EM and MM), this defualt works well. In problems where an eigenvalue of the Jacobian of $F$ is outside of the interval (0,1), step.min0 should be less than 1 or even negative in some cases.}
 \item{step.max0}{A positive-valued scalar denoting the initial value of the maximum steplength taken by a SQUAREM algorithm. Default is 1. When the steplength computed by SQUAREM exceeds step.max0, the steplength is set equal to step.max0, but then step.max0 is increased by a factor of mstep.}
 \item{mstep}{A scalar greater than 1. When the steplength computed by SQUAREM exceeds step.max0, the steplength is set equal to step.max0, but step.max0 is increased by a factor of mstep. Default is 4.}
 \item{objfn.inc}{A non-negative scalar that dictates the degree of non-montonicity. Default is 1. Set objfn.inc = 0 to obtain monotone convergence. Setting objfn.inc = Inf gives a non-monotone scheme. In-between values result in partially-monotone convergence.}
 \item{projection}{A function projecting the parameter after each iteration. Default is identity function \eqn{f(x) = x}}
 \item{tol}{A small, positive scalar that determines when iterations should be terminated, see \code{convtype} for details. Default is \code{1e-7}}
 \item{maxiter}{An integer denoting the maximum limit on the number of evaluations of \code{fixptfn}. Default is 2000.}
 \item{convtype}{A string indicating the convergence criteria.
                If it is "parameter", the algorithm will termenate when L2 norm of parameters difference \eqn{x_new - x_old < tol}.
                If it is "objfn", the algorithm will terminate when the absolute difference of objective function \eqn{|L_new - L_old| < tol}.
                If it is "user" or \code{conv.spec} is not \code{NULL}. Then the convergence is guided by the user defined function \code{conv.spec}.
                Default is "parameter".}
 \item{par.track}{An bool value indicating whether to track parameters along the algorithm. \code{TRUE} for tracking and \code{FALSE} for not. Default is \code{FALSE}}
 \item{conv.spec}{A function for user specified convergence criteria. When using "parameter" or "objfn" option in \code{convtype}, this should be \code{NULL}.
                 The function should have the form \code{f(old_parameter, new_parameter, old_objective, new_objective, tolerance)} and return 1 if convergent, 0 if not.
                 Defalut is \code{NULL}.}
}
}
\examples{
\dontrun{
set.seed(54321)
prob = lasso_task(lam=1)
squarem(prob$initfn(), prob$fixptfn, prob$objfn, X=prob$X, y=prob$y)
}

}
\references{
R Varadhan and C Roland (2008), Simple and globally convergent numerical schemes for accelerating the convergence of any EM algorithm, Scandinavian Journal of Statistics, 35:335-353.

C Roland, R Varadhan, and CE Frangakis (2007), Squared polynomial extrapolation methods with cycling: an application to the positron emission tomography problem, Numerical Algorithms, 44:159-172.

Y Du and R Varadhan (2020), SQUAREM: An R package for off-the-shelf acceleration of EM, MM, and other EM-like monotone algorithms, Journal of Statistical Software, 92(7): 1-41. <doi:10.18637/jss.v092.i07>
}
