% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_task.R
\name{tsne_task}
\alias{tsne_task}
\title{tSNE problem}
\usage{
tsne_task(perplexity = 30, rho = 2e-07, sd = 0.01)
}
\arguments{
\item{perplexity}{Perplexity value used to calculate the similarity matrix in original space before embedding}

\item{rho}{Initial Learning rate used in the MM algorithm for tSNE}

\item{sd}{Standard deviation for the random initial embedding coordinates}
}
\value{
A list containing all components needed for benchmarking the problem
 \item{initfn}{Parameter random initializing function}
 \item{fixptfn}{Updating function for the fixed point iteration problem}
 \item{objfn}{Function calculating the objective value for current parameter}
 \item{...}{Other arguments required in functions above}
}
\description{
A function that will generate the task list that can be used to benchmark the acceleration of tSNE in COIL-20 data
}
\examples{
\dontrun{
set.seed(54321)
problem = tsne_task(rho=5e-6, sd=5e-3)
benchmark(
  problem,
  algorithm=c("raw", "squarem", "daarem", "pem", "qn", "nes"),
  ntimes = 20
)
}

}
\references{
Yang Z, Peltonen J, Kaski S (2015). Majorization-minimization for manifold embedding. In: Artificial Intelligence and Statistics, 1088–1097. PMLR.
}
